refactor: Update check_model_prediction_methods to validate model methods
